feat: change directory structure
