--- bin/deptrack.ts
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-const port = process.argv[2] || 4000;
-console.log("bin test");
