--- frontend/public/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Dep Track</title>
-    <link rel="stylesheet" href="style.css" />
-  </head>
-  <body>
-    <div id="canvas">
-      <svg width="100%" height="100%"></svg>
-    </div>
-    <script type="module" src="./public/graph.js"></script>
-  </body>
-</html>
